% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss-reading.R
\name{get.cb}
\alias{get.cb}
\title{Get codebook}
\usage{
get.cb(x)
}
\arguments{
\item{x}{a S4 object returned by \code{memisc::spss.system.file}}
}
\value{
Returns a tibble data frame with variable names, label, and the coding used for the categorical variables, which include summary statistics for each catagory
}
\description{
Get information about variables from a SPSS (.sav) file
}
