% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss-reading.R
\name{show.coding}
\alias{show.coding}
\title{Return tibble with variable coding}
\usage{
show.coding(x, vars = NULL)
}
\arguments{
\item{x}{a S4 object returned by \code{memisc::spss.system.file}}

\item{vars}{string vector with the name of the variables to display the coding scheme used in the data. If \code{NULL} (default), all variables are used.}
}
\description{
The function returns a list with the coding scheme of the variables
}
